% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate against gen3.theanvil.io}
\usage{
authenticate(file)
}
\arguments{
\item{file}{character(1) file path to json credentials, as
described in the 'details' section.}
}
\value{
The bearer token used for authentication, invisibly
}
\description{
Authenticate against gen3.theanvil.io using
credentials obtained external to \emph{R}. Authentication persis for
the duration of the \emph{R} session, or until authentication
expires using criteria defined on the server.
}
\details{
To obtain credentials, visit https://gen3.theanvil.io,
login, and click on the profile icon. There you can create an
access credential as a JSON file. Download this file and
remember its location. Do not share this file with others.\preformatted{A convenient location to store the credentials file is at this
location:

    cache <- tools::R_user_dir("Gen3", "cache")
    credentials <- file.path(cache, "credentials.json")
}
}
\examples{
## Authenticate first
cache <- tools::R_user_dir("Gen3", "cache")
credentials <- file.path(cache, "credentials.json")
if (file.exists(credentials))
    authenticate(credentials)

}
