% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_stat}
\alias{download_stat}
\alias{download_object_id}
\title{Access and download objects from Gen3 google buckets}
\usage{
download_stat(object_id)

download_object_id(object_id, destination)
}
\arguments{
\item{object_id}{character(1) object identifier returned from a
query of form \code{values("sequences", "object_id", ...)}.}

\item{destination}{character(1) location to retrieve the file,
either a local path or a 'gs://' uri with appropriate write
permissions, e.g., as returned by \code{AnVIL::avbucket()} in AnVIL.}
}
\value{
\code{download_stat()} returns an object of class
\code{gcloud_sdk_result}, which when printed displays statistics on
the object.

\code{download_object_id()} returns the path or bucket in which
the object was stored.
}
\description{
\code{download_stat()} returns information about objects
available for download from Gen3.

\code{download_object_id()} retieves the object id to a
local file system or another google bucket for which the user
has write access.
}
\details{
These functions must be run in the AnVIL environment, or
with \code{gcloud} command line tools available and configured to
such that \code{AnVIL::gcloud_project()} returns a billing account
(e.g., belonging to AnVIL) that will allow 'requester pays'
downloads.\preformatted{`download_stat()` fails if the object is not available,
the user does not have permission to access the object, or the
user does not have billing enabled.
}
}
\examples{
## object_id of smallest object at time of writing; see vignette
object_id <- "dg.ANV0/76050167-ff6a-445d-9376-c3d9192fd02b"

\dontrun{
## Who will pay for the downloads?
AnVIL::gcloud_project()

download_stat(object_id)
}

\dontrun{
destination <- download_object_id(object_id, tempfile())
file.info(destination) \%>\% as_tibble()

destination <- download_object_id(object_id, AnVIL::avbucket())
destination
AnVIL::avfiles_stat(destination)
}
}
